Change km signal handling to support SA_RESETHAND and return correct exit status (#1166)
* Change km signal handling to support SA_RESETHAND and return out of band exit status

Add support for the SA_RESETHAND flag in the sa_flags field of the sigaction structure.
Also change km to send a termination signal to itself after it processes a signal that
would terminate the payload.  This allows the exit status bits from km and returned by
waitid() to be correct.  The bits of interest are the "coredumped" and "terminated by signal"
flags.  Formerly km returned the payload termination signal as km's exit status, no other flags
were set.  A parent processes would not know that the payload terminated because of an uncaught
signal.  With this change, km now coredumps when the payload coredumps.  km still produces
the kmcore file when the payload terminates and then km sends the same signal to itself
which will cause km to coredump too.  So, we will see sympathetic km coredumps which are
usually not meaningful.

Tested with the bats tests.
Also tested with the node test test/parallel/test-worker-prof.js
Fixes #1158.

* Add bats test for SA_RESETHAND flag.

* Supress the sympathetic km coredump when the payload coredumps.