Suppress km sympathetic coredumps (#1180)
* Suppress km sympathetic codedumps

This change suppresses the km coredump that happens when km arranges for its exit status
to match the payload's exit status.
Instead of using setrlimit()/prlimit() to restrict the size of the core file to zero, we use
prctl(PR_SET_DUMPABLE, SUID_DUMP_DISABLE) to disable the coredump but still return
the correct process exit status signifying the payload terminated with a signal.
prlimit() would suppress the coredump but would leave an entry in the coredumpctl
database for the empty core.
Note that SUID_DUMP_DISABLE seems to only be defined in the kernel source in
include/linux/sched/coredump.h.  I couldn't find the package containing a copy
of that source file, so we hard code this value to 0.

FYI:
Making a process not core dumpable makes the process also not ptrace()able unless the
tracing process has CAP_SYS_PTRACE capability.  Since we are setting the process not core dumpable
just before it is about to exit, I don't think this will be harmful.
We also need to be aware that marking a process as not core dumpable also makes the files
in its /proc/$pid/ directory owned by root.
"man 2 prctl", "man 5 proc", and "man 2 ptrace" give the details about these retrictions.

Tested with bats on my workstation.

Add conditional definition of SUID_DUMP_DISABLE
And a brief explanation of why we are defining it.