Save the hypercall args address across signal handler invocation. (#1187)
There seems to be a situation where a thread can make a hypercall but before kvm passes
the hypercall to km the thread is interrupted and a signal handler is scheduled to run
in that vcpu.  When the signal handler returns the pending hypercall is processed next.
If the signal handler makes a hypercall request, the address of the hypercall args overwrites
the hypercall arg address setup by the thread before the signal happened.

The fix for this is to save the hypercall args address as part of scheduling the signal
handler and then restoring the hypercall arg address when the signal handler returns.

The node test/parallel/test-worker-prof.js uncovered this problem.  Usually about 1 in
10 runs of this test would wxhibit the problem.

Tested by running the bats tests and also running the test/parallel/test-worker-prof.js
test manually.  Both tests done on my workstation.