Made nightly CI Kubernetes cluster to be cleaned up even on failures (trivial)
experience shows that we don't need this cluster in the majority of failures,
and it just adds to resource consumption and manual cleanup.

So making it cleaned up no matter what.
Behavior can be changed by setting K8S_CLUSTER_ERR_NO_CLEANUP: true

the name is a bit confusing , but it is so just to keep consistent ith other ERR_NO_CLEANUP vars

Tested with manual run of nightly