Prevent snapshots with warnings under certain circumstances (#1199)
* Prevent snapshots with warnings under certain circumstances

Since km doesn't checkpoint all needed state for a snapshoted payload to recover
km checks for these conditions and aborts the snapshot (and km) if the condition
is found.

km checks for:
- epoll fd with fd's that have pending i/o
- a pipe with in flight data
- a connected network socket
- a socketpair with in flight data
- an active setitimer() interval timer

km does not currently support create_timer() interval timers.  When it does snapshots
with active timers of this sort must be prevent the snapshot.

Fixes #1123

Added a new snapshot fail test to create conditions that should cause a snapshot failure
and then attempts a snapshot.

Tested with the bats tests on my workstation.