Fix for the node test-fs-realpath-pipe.js test
The problem was a statx() call on /proc/NNNNNNNN/fd/0 was failing when it should have worked.
The reason for the failure is proc_self_fd_name() was returning the contents of the symlink
which km_fs_statx() would pass to syscall(SYS_statx, ....).
fd 0 in this test was a socket.  And the contents of the /proc/pid/fd/0 symlink was something
like "socket:[456789]" which is not a value filename causing statx to fail.

The fix is to remove code from proc_self_fd_name() which was there to xlate the filename from
/proc/1/fd/0 to the path with the real process id in the path which would then be passed to
statx().
We also added some code to disallow attempts to access the km internal fd's above approximately
fd 727.

This was tested with the bats tests and the node "make test-all" test on my workstation.