Ensure the bats tests use known invalid pid's.
A few of the bats tests want to use an invalid pid as part of testing.
They were using 9999 which is probably invalid most of the time but sometimes
it is not.
The fix is to get the contents of /proc/sys/kernel/pid_max which is one beyond
the maximum pid the kernel supports and use that value as the invalid pid.

Tested by running the bats tests on my local workstation.
When do we start writing tests to verify that the tests are correct?
Sounds like a recursive descent into test hell.
Can't think about that, brain hurts too much.