Skip runinng the test-cluster-master-error.js and test-cluster-master-killl.js node tests
These 2 tests don't reap some of their child processes so process 1 becomes their parent.
The tests hang in ci waiting for the zombies to go away.
The problem is ci uses the tail command as process 1 for the node tests.
If you use a shell as process 1 it will reap the zombie processes.

For now, just don't run these tests.