Fix problems with running a shell as the entrypoint to a container.
If you run:

docker run -it --rm --runtime krun alpine

and then tried to run a command at the prompt, the kernel would stop the command because
it thought the command was not the owner of the controlling tty.

The basic problem was km did not let the kernel know that SIGTTOU was blocked by the payload.
When the payload would call tcsetpgrp() to take control of the tty, that would cause the payload
to be stopped by the kernel.  This is how tcsetpgrp() works.

This fix is to propagate the settings for the SIGTTOU signal to the kernel so that it can
handle this situation correctly.

A small smoke test for this problem was added to the bats signals test.