Update README.md
Updated link: User Guide link was pointing to a non-existent Google docs folder. Given the consensus to provide user documents in the repo, and given that the Google source document is now out-of-date, I made this link to user-guide.md in the km/docs folder. 

Made other minor edits to descriptive text.