Preserve SA_RESTORER address in snapshots
Snapshots were not preserving the SA_RESTORER flag but not the address.
This caused signal handlers created with the SA_RESTORER flag to attempt
to return to 0. We found this in the SpringBoot demo, but it effected more
than that.

The test excersizes the condition. Verified that the test fails without
the changes to KM.