Revert some `prog` expresessions.
The shell evalutes a command line completely before running the command line.
If a "test" expression contains a program whose output is compared there will
be a failure if the program does not exist.  For example:
[ -e /usr/bin/doesntexist && `/usr/bin/doesntexist` = "droppings" ]
will fail when the shell tries to run /usr/bin/doesntexist even thought one would think the -e in
the first expression would prevent it from running doesntexist.