Assign selinux file contexts at install time
Built files (like km) may be built on machine with or without selinux.
And, these same files may be installed on machines with or with selinux.
So, it is not important to propagate file contexts as part of building a release.
But we always install the contexts if the install target has selinux enabled.

We also need to set a file's selinux context as part of a build so that we
can test on the build machine.