vagrant box build script wasn't using the correct source code
vagrant-box.sh was using "git checkout -b ${SRC_BRANCH}"  to get the source to
produce released versions of our fedora and ubuntu boxes.
-b created an new branch with the same name as the release branch.
We really wanted to switch to the branch we want to release.

This was discovered when a new target was added to a makefile and this
script tried to use the new target but the new branch from master didn't
have the target causing the make to fail.