DOC-ONLY-fix for Docker run of kontain examples
There was a discrepancy in that the 'docker run' example would not instantiate the example programs because the language interpreter name was omitted from the command.  So either we should put the language interpreter in the base image CMD and receive args, or we have to specify it in the docs.  This patch specifies it in the docs, so folks can try out Kontain successfully.