Support KM_LOGTO environment variable.
Allows us to direct km's logging to somewhere else using an environment
variable.  We can use this to tell payloads running in a container to
trace to stderr which can appear in a container's log which we can
retrieve using "docker logs ..." or some similar command.

Added a test for this to the bats km_logging test.