WIP: Change the snapshot mgmt interface to support additional arguments
We now send a request to the km mgmt thread with the arguments for the
snapshot request.  Previously establishing a connection was enough to
trigger a snapshot.

Also changed km_cli to search for candidate processes to snapshot based
on a list of process ids and/or a list of command names.

The bats tests pass but more bats tests need to be written for the
additional new arguments to km_cli.