Add support for snapshoting data buffered in a socketpair
Also support for snapshotting data buffered in a pipe.
The buffered data is stored in the elf note for the write side
of the pipe or socketpair fd's in the snapshot.
The data in the pipe or socketpair is read out to perform the
snapshot and written back into the write side of the pipe so
the application can get its data if it were to resume after the
snapshot request returns.
Changed the bats snapshot_fail_test to now verify that data buffered
in a pipe or socketpair is now available when the snapshot is
restarted.