Change bats basic_snapshot to tolerate multiple pids from pidof
The basic_snapshot test uses the pidof command to find the pid of a
command it wants to snapshot.  When the bats tests run with concurrency
enabled pidof can return multiple pids which isn't handled.
Add a small loop to traverse the list of pids returned by pidof to find
which pid we need for the test.