Change km_dump_core() to return an error on failure. (#1681)
* Change km_dump_core() to return an error on failure.

To allow a payload to continue running after a failed snapshot request
we need to change km_dump_core to return an error instead of
aborting/exiting km.  All of the functions called by km_dump_core were
changed to return an error on failure.
Changed the callers of km_dump_core() to handle error returns.