Change snapshot to handle data buffered in half open pipes and socketpairs. (#1676)
Move buffered pipe and socketpair data to the elf note for the read end of the pipe
or socketpair.  We formerly kept the data in the elf note for the write
end of the pipe/socketpair.  With a half open pipe or socketpair the
written fd is closed after all data has been written.  The data stays
queued until the other end is ready to read it.
We can't really tell where the other end of these half open
pipes/socketpairs is.  It may still be in another process or it may have been
created local to the process and then closed.  We can snapshot the
pipe/socketpair that is local to the process being snapshotted.
We can't snapshot the data buffered in a pipe/socketpair where the other
end is in a different process since we don't support multi-process
snapshot.

So a new restriction on snapshot is being imposed.  If a process has
successfully forked, it can not be snapshotted.  This avoids dealing
with "apparently" half open pipe/socketpairs where the other end may be
in another process.

Added code to save and restore the size of data that can be buffered in
the kernel in a pipe.  No test for this yet.

Added new bats tests for snapshotting processes with half open pipes and
socketpairs and processes that have forked.

change ioctlfionread() to stop vetting passed fd for validity.

We shouldn't be passing invalid fd's to ioctlfionread() anymore with the
rest of the changes in this PR.