Change km to look for new env var KM_MGTDIR to config snapshotting.
The new environment variable KM_MGTDIR tells km where it should create
its management pipe and where it should place its snapshots when asked
to create one.
Management pipes are now created to be unique so that children of the
main km process can create pipes whose names will not collide with the
names of other pipes.  The pipename is created as follows:
  sprintf(pipename, "%s/kmpipe.%s.%d", $KM_MGTDIR,
          basename(km_payload_name), getpid());
Snapshots are place in KM_MGTDIR and are named in a similar fashion:
  sprintf(snapname, "%s/kmsanp.%s.%d", $KM_MGTDIR,
          basename(km_payload_name), getpid());

If the environment variable KM_MGTDIR exists and is non-null km will ignore the
--mgtpipe, --snapshot command line flags and the KM_MGTPIPE environment
variable.