Change gdb_lots_of_threads.c to wait for all threads to be running
I have found that threads created by pthread_create() are not running
immediately and can't be pthread_join()'ed until they are running.
This fix is being put in because the gdb_attch bats test failed because
gdb info threads reported all threads were on vcpu-0.
I don't understand how that would happen but I think those new threads
were not ready to be debugged because pthread_create() for the vcpu threads in
km were not done with their work.
Since this test does do pthread_join() we should wait for the threads to
be running before doing the join so this change doesn't hurt.