Add debug for the bats continue_after_snapshot test (#1691)
* Add debug for the bats continue_after_snapshot test

km will generate debug tracing for this test.  The trace goes into a
file.  If km fails, the trace will be printed into the bats test log so
we can see it.

In addition the km function km_add_guest_fd_internal() was exiting
gracefully when it discovered an attempt to use a file table slot that
was already in use.  So, now we abort km when this condition is found.
Add a pathetic stack tracer function