Add payload shrink capability, payload shrink trigger
Added code to have a running snapshot payload stopped and then exec to km so that
km can go wait for new payload connections before starting the snapshot
up again.
This works with the SNAP_LISTEN_PORT changes that are currently
available.

Also added a "shrink" hypercall to initiate payload shrink from the
application.  This was added for testing and can be removed or kept.

Use SOREUSEADDR in prelisten, add SNAP_LISTEN_SOCK test to bats