Ubuntu gcc 11.2 gdb_sharedlib bats test fails
The ubuntu gcc 11.2 dlopen looks like this:

Dump of assembler code for function dlopen:
   0x00000000002625bc <+0>:     endbr64
   0x00000000002625c0 <+4>:     push   %r15
   0x00000000002625c2 <+6>:     push   %r14
   0x00000000002625c4 <+8>:     push   %r13
   0x00000000002625c6 <+10>:    push   %r12
   0x00000000002625c8 <+12>:    push   %rbp
   0x00000000002625c9 <+13>:    push   %rbx
   0x00000000002625ca <+14>:    sub    $0x1b8,%rsp
   0x00000000002625d1 <+21>:    mov    0x4f918(%rip),%rax        # 0x2b1ef0 <head>
   0x00000000002625d8 <+28>:    mov    %rdi,0x20(%rsp)
   0x00000000002625dd <+33>:    mov    %esi,0x14(%rsp)
   0x00000000002625e1 <+37>:    movq   $0x0,0x60(%rsp)
   0x00000000002625ea <+46>:    test   %rdi,%rdi
   0x00000000002625ed <+49>:    je     0x262d75 <dlopen+1977>

We need to get the head of the dynamically loaded library list from the
instruction at offset +21.
This is a new offset for us so it needs to be added to the lib list
walker code in km_link_map_walk().
In addition if we can't find the head of the library list we need to
fail the gdb_sharedlib bats test with a clearer message.  If we see this
message:
   "Can't find the head of the head of payload's loaded library list"
we will know that dlopen() has changed and we need to look at it.