Give better error when SNAP_PORT_LISTEN env var can't be decoded
light_snap_listen() decodes what is stored in the SNAP_LISTEN_PORT env
var.  If it can't decide the variable, it just drops thru to the listen
call with an invalid fd which causes listen() to fail.
This change adds code to print a message showing the contents of
SNAP_LISTEN_PORT before exiting km.