Add support for vex instruction prefix to km_decode.c
Addeded support for the VEX 2 and 3 byte instruction prefixes.
Also added a smoke test for these prefixes to the bats decode_test.
Since the XOP prefix is similar to the VEX 3 byte prefix, it is plumbed
in but its tables of opcodes that reference memory has not been filled
in.