Improve snapshot messages and uninterpreted opcode message
Todays CI failure shows a km trace that has both a snapshot generation
and snapshot resume all run together.  To help us realize what is
happening the next time we see this there are km traces now for begin
and end of snapshot generation and resumption.

Plus The message for an uninterpreted opcode has been changed to include
both the PC for the instruction and the first 6 bytes of the instruction.