Create snapshots with holes where unbacked pages in the dumped memory
To create a snapshot with unbacked memory pages put into the core file
as holes, define SPARSE_COREFILE in the km environment when starting up
km.  Currently we only produce sparse snapshot files.  It is easy to
produce sparse corefiles if we think we need to.

This commit is mostly for people to look at the code.  There is no bats
test yet.  There is some code added to the multilisten test to allocate
100MB with malloc() and 100MB with mmap().  malloc() creates backed
pages, mmap() creates unbacked memory pages.