Add the aio io_*() syscalls to km and a simple test program
The io_setup(), io_submit(), io_getevents(), io_cancel(), and
io_destroy() hypercalls are added to km with this change.
The bats aio_test program is added to exercise the hypercalls.
I can't figure out how to make io_cancel() work.  It seems like the io
request being canceled must be active.  If the io has completed, I
assume it is nolonger cancelable and it returns an error code that
doesn't tell me the io request has completed.

We prevent snapshot when there are active asynch i/o contexts.
This is because the kernel assigns io context id's which the
payload remembers and uses in other io_*() syscalls.  We can't take a
context id away from the payload and give it a substitute when we resume
a snapshot so we disallow snapshot.  Evfentaully we should add a table
that maps between io contexts the payload knows about and io contexts
create when the snapshot is resumed.

The aio_test does a simple test of the io_*() system calls.
The io cancel part of the test is not yet trustworthy since it fails
with the error EINVAL when the supplied io context is valid.  The man
page for io_cancel claims EINVAL means invalid io context argument.
To ope with this the io_cancel() part of aio_test first finds out it there are
any i/o requests still running.  If not we don't have anything to cancel
so we don't attempt the cancel.  If requests are running we pick one and
cancel it.  io_cancel() seems to fail with the error EINVAL even when a
valid io context argument is supplied (contrary to what the man page
claims).  I assume EINVAL means the io we are trying to cancel is
unknown because it completed.  So io_cancel() returning EINVAL is not
considered a failure.
If we ever find out how io_cancel() really works we can make that part
of the test better.  Yes, I have looked at the kernel code and it does
seem to return EINVAL if the request being canceled is not found.