bats aio_test needs to run km_with_timeout in subshell with valgrind
The km_with_timeout() shell function keeps adding things to the KM_ARGS
shell variable if we don't run each instance in a subshell with the bats
run command.  When running bats with valgrind this will add
"--copyenv=LD_PRELOAD" to KM_ARGS each time km_with_timeout is run in
the main shell which will cause km to fail complaining about multiple
instances of the --copyenv flag.

I'm not sure if this is a bug in tests/test_helper.bash (which is where
km_with_timeout lives) or a bug in tests/km_core_tests.bats since we
don't consistently use run with km_with_timeout.