Log payload fd when recovered file has mismatched file mode (#1745)
* Log payload fd when recovered file has mismatched file mode

We occasionally see file mode mismatches when a payload snapshot is
resumed.  This causes the snapshot resume to fail.
In one case the original file is a regular file, after the snapshot is
resumed the file is a char special file.
I'm assuming the fd is one of the stdio fd's.  If so, we can skip this
check.
For now, add the fd to the message to see if my guess is correct.