Use restored machine.brk value during snapshot restore
Some code in snapshot restore uses the hardcoded value 256 * GIB to
decide if it is dealing with memory allocated with the brk() system call
or memory allocated with mmap().
It turns out machine.brk is saved in the coredump/snapshot and is
restored when a snapshot is recovered, so we can use that value to make
decisions rather than the hardcoded 256gb value during the remainder of
snapshot restore.