Change bats continue_after_snapshot test to discover if html has timed out (#1773)
In the wonderful cloud environment our tests are treated to all manner of
process scheduling preemptions.
In this case I thing the simple html server in the test was timed out
by km_with_timeout before we finished the loop where we take snapshots
and run curl against the html server.
We try to detect that km_cli or curl failed because the html server
process timed out.  If that is what happened we end the test but call it
successul.
If the html srver is still running then we assume there was some other
failure and we do fail the test.

We will just need to wait for mr cloud to grace us with more kookie
scheduling preemptions to see if this change helps.