Try to fix th ci failure in the bats gdb_nextstep test
We've been seeing this failure in the gdb_netstep test lately.

The only recent change "seemingly" related to this was a change where we
save the value of errno for use later in km_vcpu_one_kvm_run().
That change shifted the setting of state to IN_GUEST one statement
further away from the ioctl( KVM_RUN ) call.

This change is more an experiment to see if it "helps" fix the assertion
failure.  If it does "fix" the problem, we will know there is a race here that
we've only made less likely but probably not fixed.