Fix poll() and ppoll() hypercalls to detect snapshot'ed network socket errors
km_fs_poll() was using the incorrect value for the fd when checking for
disconnected connections in a resumed snapshot().  This would cause an
incorrect error to be returned.

Also added a fix to ppoll() where a NULL fds pointer should be allowed
under some conditions.  This change is from some arm work Srini is
doing.

This also adds the ppoll() hypercall to km and a simple ppoll test
program.